Move the functions for placing the beacon on the same file
